tcpprobe: leverage libbpf_compatible to select embedded BTF
The field `custom_btf_paths` in `struct bpf_prog_s` is actually
redundant, as the custom BTF paths are already available through `struct
bpf_object` in the skeleton. However, fields of `struct bpf_object`
resides in `libbpf.c` instead of header files and cannot be considered
as part of the public interface, therefore not able to be referenced.