init_probes: check state of listening cmd server carefully
if gala-gopher exited abnormally, such as killed by sigkill, the cmd
server sock file will remain. When systemd recovers gala-gopher and
executes init_probes.sh, it will pass the check of cmd server directly
but the listening sock may not be ready. So check the sock state instead.

Also, change the directory of pid file to the same of sock file.